requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.0
django-cors-headers==4.3.0
pandas==2.1.4 # For CSV processing